pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Artem Beletzky <belezky.artem@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.hexlet_code:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
pytest = "^7.1.3"

[tool.black]
line-length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
